Add a main-go-service to replace main-go
The new workflow does not rely on GoReleaser to build the Docker images. This has tradeoffs: the GoReleaser configuration is simplified and does not rely on passing the --push or --load flags to buildx build to bypass snapshot releases; we also get more consistent metadata with the other images; and we get improved CI build caching. The downside is that building a local Docker image is less trivial.